Cornerstone/requirements.txt
flask==0.10.1
cassandra-driver==2.1.4
blist==1.3.6
